# Speed up git diff highlight generation