@KN4CK3R 2021-06-17
Maybe we could get rid of the cache later by getting a lexer for the diff file and pass it to the `Code` method instead of asking for a lexer all the time.