@typeless 2021-06-17
@KN4CK3R Yeah, I thought about that as well.  But I guess the original author wanted to separate the `chroma` dependency from the upper layer, so I didn't do that (The change would be more invasive too).