@KN4CK3R 2021-06-17
> chdir /drone/src/modules/git/tests/repos/repo1_TestRepository_GetTag: no such file or directory

I don't think so.